requirements.txt
titanoboa==0.1.8